Fix raw input for WOW64 (32bit process on x64 OS)
On WOW64 the union offset is 8 byte aligned, so not the same as the structures definition.
Meaning we have to work out the correct offset in WOW64, because of this I've made ProcessInput take a data pointer, which has been corrected.
So the the device can cast it to the respective device type.