Re-factor some model related things.
There is not a interface for models IModel, that is ref counted.
The XModel implementation is now inside the 3DEngine lib.
The need for a Model Loader class has been removed, model loading code is not part of the XModel implementation in 3DEngine. (removes need for friend also <3)
Model manger has been updated to return IModel instead of void now I've decided what it's going to return.