Have the Editor manager not take file names but AssetNames, if the editor knows it's dealing with asset names and types it make the logic more nice.
since if the editor thinks it's only dealing with files on disk yet everything is actually coming from the db
it makes the logic a bit awkward.

Also pretty much everything this program will ever display is going to be from assetDb so might have it be the main logic
and add in support for disk files later if ever required.