Actuall I forgot that CreateDirectorytree was already removing filenames.
So code is actually relying on that.
Also it makes the call sites for alot of code quite messy if they have to remove file path.
And we already have a temp buffer in the filesystem to remove the filename from so it makes more snese to remove it there.

Also remove the warnings about directoryExists called on path with filename, as the OS code will log a error.

I'm leaving the check on deleteDirectory tho.