Remove the include of IFileSys.h in ISerialize as it only needs a forward declare.
Side effect of this is lots of files where incorrectly relying on this to get the defs from IFileSys.h so I've had to include it in a number of files.

Also remove a win32 specific include from FileSys.h <io.h>
It's needed by PathUtil and XFindData only currently.
PathUtil is 32specific so it can include it.
XFindData needs to be updated a bit to become none platform dependant.