Start to work out a system for dealing with render state.
This is sort of a first draft which requires all state to be created before draw time.
So that the render loop is very very simple, requiring only a int compare to know if state has changed.
And won't have to stall for new PSO's.

If I stick with this design i'll throw in pools and better state creation for performance but no point if I'm just gonna change it all.