update the PrimContext to use pages, so it can grow it's vertexBuffer requirements slowly.
So we can create a resonably size vertexBuffer and when it's full we move to another instead of having to resize.
This means the prim context can adapat better to each use case a instance is been used in.

Lazy loading of vertex buffers is also now done, so now if the console is never opened a vertexBuffer is never created.

Since we can limit verts per a page without it been enforced as a global limit BuffersEntry's can store 16bit offsets and counts saving 4bytes.