Support ref counting the cb handles and add a destoryCB.
We can use the ref count to know when to release the device instance.
This logic is kinda ugly maybe for cbuffers it's better if render sytem handles ref counting.
Since the render system if fully aware of Cbuffer type.

Would mean the ref count could go in same memroy block.