Start to switch to just getting the primative materials for PrimDrawing.
This then means we don't have to define the state in code for creating a primative material we can just load the material.
Which means I don't have to duplicate logic I can just use the offline pipeline for creating materials with the added benfit of the states then been data driven.