Add initial implementation of a TechDefStateManager, which handles the creation of techStates.
So materials will just ask for it's techDefState from the manager and get back all the techs for that tech def along with the tech and permatations for it.
The key thing is that they are shared, so 300 materials all using the same techDef type will all get given same techDefState from the cache.

So the render states and hardware shaders are only requested from the render system once.