Start to add in support for defining tech params as well as ability to set some ui properties.
This will mean the AssetManager will know what params are required to be shown for the current material type.
This also lets the material compiler know what props need to be stored in the materials file.

There is also a concept of boundTextures, which is just letting me state that a resource from the pixelShader stage is defined by the code