Make it so raw_file's are stored on disk with the hash after the name.
This is first step into keeping history of raw_files, in that we don't override the raw_file on disk when it's updated.
Currently they are just left dangling there, but they could be manually restored if required.

Next step will be to keep the raw_file table row and just make new ones, so we can provide api for listing raw_file api.