Add some temp logic that detects if the new asset ref is actually the current one but it's just been renamed.
This is due to the fact when I rename a asset I don't update the values props for any assets that ref it.
I think a better way to handle this tho is if props don't store the name but the id, that way we don't have to do anything id we name a asset as nothing will be stale.