Make it so we can create a scene then later add it for active simulation.
This is so we can create the scene and populate it with all the static geo over multiple frames before starting simulation on it.
It also means we can switch out the active simulation scene without having to recreate the scene.

Some refactor of the steppers would also allow multiple active scenes to be supported.