Make it so contunations are only run when all children jobs are also done, instead of running them soon as the job's function has finished or any of it's children.
This now means a continations can't be a child of a job, otherwise we will deadlock.
I've added a assert to check one level up, but it won't catch deeper childen.

I've also made it so continuations are run before telling the parent we are done, may change this later.