Expose the thread policy and don't lock in find / create asset both these calls need to be performed inside a single lock scope.
So currently i'm shifting the lock responsibilities to the caller, but I may change it if I can thing of a better way.
Also I think it would be good if we supported read/write locks, since we only need to read to find then promote to write on create.