Remove the createRenderTarget api and replace it with pixelBuffer api. Seperate functions for creating Color and depth buffers tho, since for depth buffer specifying format is not allowed per instance, so sharing same function for creating both types of buffers makes the api less clear.
This implementation is just first draft and still needs polishing / finishing.

The creating and releasing of the Pixelbuffers is been delegated to the texturemanager as i'm trying to make the concept of these complex buffers more transparent so binding
them in shader stages is same as passing a normal texture id.