Remove a lot of logic from InstancedPage, it's logic we don't need per page, but just once when submitting.
Make it so pages are fixed size and not requested that way the logic is nicer for submitting.

Added a ShpaeInstanceContainer that stored all the shape draw request for a given shape.
It also counts how many of each lod / solid or not where requested, so that post sort we know the lod ranges in buffer without having to check.
This means we have less seperate buffers and must sort, but results in less gpu buffer copyies which should pay for the sort.