Make the logic a bit more clear for file node offset / size.
I was un sure if the taking of fileHeader from size was correct,
but of course it was since we had shiffted the offset by sizeof(FileHeader)
making the size bigger by sizeof(FileHeader) as size was calculated from real file offset.
I've changed the logic to be more easy to reason about it's correctness.